add tags to node groups

+++ 97-delete-applications.sh
@@ -13,7 +13,12 @@ apps=$(kubectl get application -n argocd -o name | xargs -I{} basename {})
 for app in $apps
 do
     kubectl patch app $app  -p '{"metadata": {"finalizers": ["resources-finalizer.argocd.argoproj.io"]}}' --type merge -n argocd
-    kubectl delete app $app -n argocd --wait=false
+    if [[ "$app" == "aws-ebs-csi-driver" ]]
+    then
+        echo "Skipping aws-ebs-csi-driver"
+    else
+        kubectl delete app $app -n argocd --wait=false
+    fi
 done
 
 kubectl delete -k ../renovate-tekton-argo-talk/k8s/argocd
